refactor: reorganize analytics module into separate package structure
