+++ src/main/java/com/DevSprint/voluntrix_backend/analytics/controllers/OrganizationAnalyticsController.java
@@ -1,4 +1,4 @@
-package com.DevSprint.voluntrix_backend.controllers;
+package com.DevSprint.voluntrix_backend.analytics.controllers;
 
 import java.util.List;
 
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.DevSprint.voluntrix_backend.dtos.analytics.MonthlyDonationData;
-import com.DevSprint.voluntrix_backend.services.OrganizationPaymentAnalyticsService;
+import com.DevSprint.voluntrix_backend.analytics.dto.MonthlyDonationData;
+import com.DevSprint.voluntrix_backend.analytics.services.OrganizationPaymentAnalyticsService;
 
 @RestController
 @RequestMapping("/api/analytics/organizations")
